Feature - Create New Project Page, Bug Fix & Update
- Bug Fix useDocumentOnce
After useDocumentOnce mounts, hook that sets status to pending only once
after mount.
Afterwards when user_id or collection changes, status is on complete |
error.
Hook that fetches checks if status is pending (is not) and won't fetch
new data.
Hook that triggers fetch, needs to set status to pending when user_id or
collection is modified.

- Dropdown allows user to visit his own profile and closes if user
  clicks to visit profile.

- Bug Fix - App component
App component rendered routes when user status is idle (at start).
This led to rendering components in state where we don't know if user is
signed in or not.